upgrade: prisma v7 and fix conflict
