+++ diagram/warehouse-management-system/warehouse-management-system-part-1-2025-12-25-154102.mmd
@@ -1,0 +1,36 @@
+sequenceDiagram
+        actor MAN as Manager
+        participant FE as FrontEnd
+        participant BE as BackEnd
+        participant DB as Databse
+        actor SEL as Seller
+        autonumber
+        par CORD from Manager to Databse
+            MAN->>FE: CORD with warehouse
+            FE->>BE:(POST) /warehouse
+            BE->>DB: Create new warehouse record
+            FE->>BE:(PUT) /warehouse
+            BE->>DB: Update warehouse record
+            FE->>BE:(DELETE) /warehouse
+            BE->>DB: Delete warehouse record
+        end
+        par Get and Create new order
+            SEL->>FE: Get warehouse's list
+            FE->>BE:(GET) /warehosue
+            BE->>DB: Get warehouse's list
+            DB-->>FE: Return Data
+            SEL->>FE: Create new order
+            FE->>BE: (POST) /order-warehosue
+            BE->>DB: Create new order warehouse record
+        end
+        BE->>MAN: Send message to Manager
+        par CORD with order warehouse
+            MAN->>FE:Get order list
+            FE->>BE: (GET) /order-warehouse
+            BE->>DB: Get order warehouse list
+            DB-->>FE: Return Data
+            MAN->>FE: Change order status
+            FE->>BE: (POST) /order-warehouse
+            BE->>DB: Update order status
+        end
+        BE->>SEL: Send message to seller
