+++ next.config.ts
@@ -3,6 +3,12 @@ import createNextIntlPlugin from 'next-intl/plugin';
 const withNextIntl = createNextIntlPlugin();
 
 const nextConfig: NextConfig = {
+  serverExternalPackages: [
+    '@prisma/client',
+    '@prisma/adapter-pg',
+    'pg',
+    '@prisma/client-runtime-utils',
+  ],
   images: {
     remotePatterns: [
       {
