+++ package.json
@@ -26,6 +26,8 @@
     "@google/genai": "^1.31.0",
     "@hookform/resolvers": "^5.2.2",
     "@icons-pack/react-simple-icons": "^13.8.0",
+    "@prisma/adapter-pg": "^7.2.0",
+    "@prisma/client-runtime-utils": "^7.2.0",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-checkbox": "^1.3.3",
@@ -56,6 +58,7 @@
     "cmdk": "^1.1.1",
     "date-fns": "^4.1.0",
     "dayjs": "^1.11.18",
+    "dotenv": "^17.2.3",
     "ecommerce_nextjs": "link:",
     "effect": "^3.19.6",
     "embla-carousel-autoplay": "^8.6.0",
@@ -67,6 +70,7 @@
     "next-intl": "^4.3.11",
     "next-themes": "^0.4.6",
     "nodemailer": "^7.0.10",
+    "pg": "^8.16.3",
     "pusher": "^5.2.0",
     "pusher-js": "^8.4.0",
     "qs": "^6.14.0",
@@ -80,6 +84,7 @@
     "recharts": "2.15.4",
     "redis": "^5.10.0",
     "resend": "4.7.0",
+    "server-only": "^0.0.1",
     "sonner": "^2.0.7",
     "stripe": "^19.3.0",
     "swiper": "^12.0.3",
@@ -95,10 +100,12 @@
     "@eslint/eslintrc": "^3",
     "@eslint/js": "^9.39.1",
     "@next/eslint-plugin-next": "16.0.8",
-    "@prisma/client": "6.19.0",
+    "@prisma/client": "7.2.0",
+    "@prisma/config": "^7.2.0",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
     "@types/nodemailer": "^7.0.4",
+    "@types/pg": "^8.16.0",
     "@types/qs": "^6.14.0",
     "@types/react": "19.2.7",
     "@types/react-dom": "19.2.3",
@@ -117,7 +124,7 @@
     "pnpm": "^10.22.0",
     "postcss": "^8.5.6",
     "prettier": "3.6.2",
-    "prisma": "^6.19.0",
+    "prisma": "^7.2.0",
     "tailwindcss": "^4",
     "ts-node": "^10.9.2",
     "tsx": "^4.20.6",
