+++ prisma/schema/auth.prisma
@@ -1,0 +1,33 @@
+//================== Prisma Schema =================//
+// File: prisma/auth.prisma
+// Schema for Authentication models in Prisma
+//==================================================//
+
+model Session {
+    id        String   @id @default(uuid()) @db.Uuid
+    expiresAt DateTime
+    token     String   @unique
+    createdAt DateTime @default(now())
+    updatedAt DateTime @updatedAt
+    ipAddress String?
+    userAgent String?
+    userId    String   @db.Uuid
+    user      User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+
+    impersonatedBy String?
+
+    @@index([userId])
+    @@map("session")
+}
+
+model Verification {
+    id         String   @id @default(uuid()) @db.Uuid
+    identifier String
+    value      String
+    expiresAt  DateTime
+    createdAt  DateTime @default(now())
+    updatedAt  DateTime @default(now()) @updatedAt
+
+    @@index([identifier])
+    @@map("verification")
+}
