+++ prisma/schema/cart.prisma
@@ -1,0 +1,32 @@
+//================== Prisma Schema =================//
+// File: prisma/cart.prisma
+// Schema for Cart models in Prisma
+//==================================================//
+
+model Cart {
+    id        String   @id @default(uuid()) @db.Uuid
+    userId    String   @unique @db.Uuid
+    user      User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+    createdAt DateTime @default(now())
+    updatedAt DateTime @default(now()) @updatedAt
+
+    items CartItem[]
+
+    @@map("cart")
+}
+
+model CartItem {
+    id        String   @id @default(uuid()) @db.Uuid
+    cartId    String   @db.Uuid
+    variantId String   @db.Uuid
+    quantity  Int      @default(1)
+    priceSnap Decimal  @db.Decimal(14, 2)
+    currency  Currency @default(VND)
+    addedAt   DateTime @default(now())
+
+    cart    Cart           @relation(fields: [cartId], references: [id], onDelete: Cascade)
+    variant ProductVariant @relation(fields: [variantId], references: [id], onDelete: Restrict)
+
+    @@unique([cartId, variantId])
+    @@map("cart_item")
+}
