+++ prisma/schema/payout.prisma
@@ -1,0 +1,114 @@
+//================== Prisma Schema =================//
+// File: prisma/payout.prisma
+// Schema for Payout model in Prisma
+//==================================================//
+
+//=================== ENUMS ===================//
+enum LedgerDirection {
+    CREDIT // Cộng tiền vào (+/Tăng)
+    DEBIT // Trừ tiền ra (-/Giảm)
+}
+
+enum WalletType {
+    SHOP // Ví của người bán
+    USER // Ví của người mua
+    PLATFORM // Ví của sàn (thu phí sàn, giữ tiền hộ)
+}
+
+enum LedgerType {
+    ORDER_PAID // khách trả tiền
+    VOUCHER_SUBSIDY // sàn bù tiền
+    ADJUSTMENT_IN // chỉnh tay +
+    PLATFORM_FEE // phí sàn
+    SHIPPING_FEE // phí ship
+    PAYOUT
+    SETTLEMENT // pending -> available
+    REFUND // hoàn tiền cho khách
+    ADJUSTMENT_OUT // chỉnh tay -
+}
+
+enum IntentStatus {
+    ACTIVE // đang chờ thanh toán
+    SUCCEEDED // thanh toán thành công
+    FAILED // gateway báo fail
+    EXPIRED // quá hạn
+    CANCELLED // user hủy
+}
+
+enum SettleStatus {
+    PENDING
+    PROCESSED
+    CANCELLED
+}
+
+enum PayoutStatus {
+    REQUESTED
+    PROCESSING
+    COMPLETED
+    REJECTED
+}
+
+//=================== MODELS ===================//
+model LedgerEntry {
+    id            String          @id @default(uuid()) @db.Uuid
+    walletId      String          @db.Uuid
+    walletType    WalletType      @default(SHOP)
+    amount        Decimal         @db.Decimal(19, 4)
+    direction     LedgerDirection // CREDIT hoặc DEBIT
+    balanceBefore Decimal         @default(0) @db.Decimal(19, 4)
+    balanceAfter  Decimal         @default(0) @db.Decimal(19, 4)
+    type          LedgerType
+    description   String?
+
+    transactionGroup String?
+    paymentId        String? @db.Uuid
+    orderId          String? @db.Uuid
+
+    createdAt DateTime @default(now())
+    Shop      Shop?    @relation(fields: [shopId], references: [id])
+    shopId    String?  @db.Uuid
+
+    @@index([walletId, createdAt])
+    @@index([type])
+    @@index([orderId])
+    @@index([paymentId])
+    @@index([transactionGroup])
+}
+
+model SettlementQueue {
+    id      String  @id @default(uuid())
+    shopId  String
+    orderId String  @unique // 1 đơn 1 lần trả
+    amount  Decimal @db.Decimal(19, 4)
+
+    feeAmount Decimal @default(0) @db.Decimal(19, 4)
+    netAmount Decimal @default(0) @db.Decimal(19, 4) // Số thực nhận = amount - fee
+
+    status SettleStatus // PENDING, PROCESSED, CANCELLED
+    dueAt  DateTime // Ngày đến hạn (settleDate)
+
+    retryCount Int     @default(0)
+    failReason String?
+
+    createdAt DateTime @default(now())
+    updatedAt DateTime @updatedAt
+
+    @@index([status, dueAt])
+    @@map("settlement_queue")
+}
+
+model PayoutRequest {
+    id     String       @id @default(uuid())
+    shopId String       @db.Uuid
+    amount Decimal      @db.Decimal(19, 4)
+    status PayoutStatus // REQUESTED, PROCESSING, COMPLETED, REJECTED
+
+    bankInfo Json // Thông tin TK ngân hàng nhận tiền
+    txRef    String? // Mã tham chiếu chuyển khoản ngân hàng
+
+    createdAt DateTime @default(now())
+    updatedAt DateTime @updatedAt
+
+    @@index([shopId, status])
+    @@map("payout_request")
+}
