+++ prisma/schema/product.prisma
@@ -1,0 +1,134 @@
+//================== Prisma Schema =================//
+// File: prisma/product.prisma
+// Schema for Product model in Prisma
+//==================================================//
+
+//=================== ENUMS ===================//
+enum ProductStatus {
+    DRAFT
+    PUBLISHED
+    ARCHIVED
+}
+
+enum Visibility {
+    PUBLIC
+    UNLISTED
+    PRIVATE
+}
+
+enum Currency {
+    VND
+    USD
+}
+
+//=================== MODELS ===================//
+model Product {
+    id          String        @id @default(uuid()) @db.Uuid
+    shopId      String        @db.Uuid
+    shop        Shop          @relation(fields: [shopId], references: [id], onDelete: Cascade)
+    categoryId  String?       @db.Uuid
+    category    Category?     @relation(fields: [categoryId], references: [id], onDelete: SetNull)
+    title       String
+    slug        String        @unique
+    origin      String?
+    description String?
+    status      ProductStatus @default(DRAFT)
+    visibility  Visibility    @default(PUBLIC)
+    attributes  Json?
+    ratingAvg   Decimal       @default(0) @db.Decimal(2, 1)
+    ratingCount Int           @default(0)
+    soldCount   Int           @default(0)
+    minPrice    Decimal       @db.Decimal(14, 2)
+    maxPrice    Decimal       @db.Decimal(14, 2)
+    currency    Currency      @default(VND)
+    createdAt   DateTime      @default(now())
+    updatedAt   DateTime      @default(now()) @updatedAt
+    deletedAt   DateTime?
+
+    images   ProductImage[]
+    variants ProductVariant[]
+    // system tags
+    tags     Tag[]
+
+    keywords String[]
+
+    reviews        Review[]
+    questions      ProductQuestion[]
+    wishlisted     WishlistItem[]
+    orderItems     OrderItem[]
+    VoucherProduct VoucherProduct[]
+    Message        Message[]
+
+    @@index([shopId, status, visibility])
+    @@index([categoryId])
+    @@map("product")
+}
+
+model ProductImage {
+    id        String   @id @default(uuid()) @db.Uuid
+    productId String   @db.Uuid
+    product   Product  @relation(fields: [productId], references: [id], onDelete: Cascade)
+    url       String
+    publicId  String?
+    alt       String?
+    position  Int      @default(0)
+    createdAt DateTime @default(now())
+
+    @@index([productId, position])
+    @@map("product_image")
+}
+
+model ProductVariant {
+    id            String   @id @default(uuid()) @db.Uuid
+    productId     String   @db.Uuid
+    product       Product  @relation(fields: [productId], references: [id], onDelete: Cascade)
+    sku           String   @unique
+    name          String?
+    price         Decimal  @db.Decimal(14, 2)
+    image         String
+    imagePublicId String?
+    compareAt     Decimal? @db.Decimal(14, 2)
+    currency      Currency @default(VND)
+    stock         Int      @default(0)
+    reserved      Int      @default(0)
+    weightGrams   Int?
+    lengthMm      Int?
+    widthMm       Int?
+    heightMm      Int?
+    attributes    Json?
+    isActive      Boolean  @default(true)
+    createdAt     DateTime @default(now())
+    updatedAt     DateTime @default(now()) @updatedAt
+
+    orderItems OrderItem[]
+    cartItems  CartItem[]
+
+    @@index([productId, isActive])
+    @@map("product_variant")
+}
+
+model ProductQuestion {
+    id         String    @id @default(uuid()) @db.Uuid
+    productId  String    @db.Uuid
+    userId     String    @db.Uuid
+    body       String
+    answer     String?
+    answeredAt DateTime?
+    createdAt  DateTime  @default(now())
+
+    product Product @relation(fields: [productId], references: [id], onDelete: Cascade)
+    user    User    @relation(fields: [userId], references: [id], onDelete: Cascade)
+
+    @@index([productId])
+    @@map("product_question")
+}
+
+model Tag {
+    id   String @id @default(uuid()) @db.Uuid
+    name String
+    slug String @unique
+
+    products Product[]
+
+    @@map("tag")
+}
