+++ prisma/schema/schema.prisma
@@ -1,0 +1,10 @@
+generator client {
+  provider      = "prisma-client-js"
+  output        = "../../lib/generated/prisma"
+  binaryTargets = ["native", "rhel-openssl-3.0.x"]
+}
+
+datasource db {
+  provider = "postgresql"
+  // url      = env("DATABASE_URL")
+}
