+++ prisma/schema/warehouse.prisma
@@ -1,0 +1,99 @@
+//================== Prisma Schema =================//
+// File: prisma/warehouse.prisma
+// Schema for Warehouse model in Prisma
+//==================================================//
+
+//=================== ENUMS ===================//
+enum WarehouseStatus {
+    OPEN
+    CLOSED
+    UNDER_MAINTENANCE
+    FULL
+}
+
+enum StorageAreaType {
+    COLD_STORAGE
+    DRY_STORAGE
+    HAZARDOUS_MATERIALS
+    GENERAL_STORAGE
+    ELECTRICAL_EQUIPMENT
+}
+
+enum SlotStatus {
+    AVAILABLE
+    OCCUPIED
+    RESERVED
+    DAMAGED
+}
+
+//==================== MODELS ===================//
+model Warehouse {
+    id        String          @id @default(uuid()) @db.Uuid
+    name      String
+    location  String
+    metadata  Json?
+    status    WarehouseStatus @default(OPEN)
+    createdAt DateTime        @default(now())
+    updatedAt DateTime        @updatedAt
+
+    //Relations
+    areas StorageArea[]
+
+    @@index([status, id])
+    @@map("warehouses")
+}
+
+model StorageArea {
+    id          String          @id @default(uuid()) @db.Uuid
+    name        String
+    type        StorageAreaType @default(GENERAL_STORAGE)
+    warehouseId String          @db.Uuid
+    status      WarehouseStatus @default(OPEN)
+    createdAt   DateTime        @default(now())
+    updatedAt   DateTime        @updatedAt
+
+    //Relations
+    warehouse Warehouse @relation(fields: [warehouseId], references: [id], onDelete: Cascade)
+    slots     Slot[]
+
+    @@index([warehouseId, type])
+    @@map("storage_areas")
+}
+
+model Slot {
+    id        String     @id @default(uuid()) @db.Uuid
+    areaId    String     @db.Uuid
+    serial    String
+    productId String?    @db.Uuid
+    ownerId   String?    @db.Uuid
+    status    SlotStatus @default(AVAILABLE)
+    createdAt DateTime   @default(now())
+    updatedAt DateTime   @updatedAt
+
+    //Relations
+    area       StorageArea @relation(fields: [areaId], references: [id], onDelete: Cascade)
+    shop       Shop?       @relation(fields: [ownerId], references: [id])
+    orderSlots OrderSlot[]
+
+    @@index([id, status])
+    @@map("slots")
+}
+
+//===================
+//Store Order to use warehouse slots
+//===================
+model OrderSlot {
+    id        String   @id @default(uuid()) @db.Uuid
+    orderId   String   @db.Uuid
+    slotId    String   @db.Uuid
+    quantity  Int
+    createdAt DateTime @default(now())
+    updatedAt DateTime @updatedAt
+
+    //Relations
+    slot  Slot  @relation(fields: [slotId], references: [id], onDelete: Cascade)
+    order Order @relation(fields: [orderId], references: [id], onDelete: Cascade)
+
+    @@index([orderId, slotId])
+    @@map("order_slots")
+}
