+++ prisma/schema/wishlist.prisma
@@ -1,0 +1,28 @@
+//================== Prisma Schema =================//
+// File: prisma/wishlist.prisma
+// Schema for Wishlist models in Prisma
+//==================================================//
+
+model Wishlist {
+    id        String   @id @default(uuid()) @db.Uuid
+    userId    String   @unique @db.Uuid
+    user      User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+    createdAt DateTime @default(now())
+
+    items WishlistItem[]
+
+    @@map("wishlist")
+}
+
+model WishlistItem {
+    id         String   @id @default(uuid()) @db.Uuid
+    wishlistId String   @db.Uuid
+    productId  String   @db.Uuid
+    createdAt  DateTime @default(now())
+
+    wishlist Wishlist @relation(fields: [wishlistId], references: [id], onDelete: Cascade)
+    product  Product  @relation(fields: [productId], references: [id], onDelete: Cascade)
+
+    @@unique([wishlistId, productId])
+    @@map("wishlist_item")
+}
